all bonuses

+++ routes/courses.routes.js
@@ -17,40 +17,62 @@ router.get('/', (req, res, next) => {
 })
 
 router.get('/create', async (req, res, next) => {
-    try {
-        const allstudents = await UserModel.find({ role: STUDENT }).select('username _id')
-        const devs = await UserModel.find({ role: DEV }).select('username _id')
-        const teachers = await UserModel.find({ role: TA }).select('username _id')
-        res.render('courses/create-course', { allstudents, teachers, devs })
+    if (req.app.locals.isTa) {
+        try {
+            const allstudents = await UserModel.find({ role: STUDENT }).select('username _id')
+            const devs = await UserModel.find({ role: DEV }).select('username _id')
+            const teachers = await UserModel.find({ role: TA }).select('username _id')
+            res.render('courses/create-course', { allstudents, teachers, devs })
 
-    } catch (err) {
-        console.log(err)
-        next(err)
+        } catch (err) {
+            console.log(err)
+            next(err)
+        }
+    } else {
+        res.render('auth/login', { errorMessage: 'No tenes premiso weon' })
     }
-
 })
 router.get('/:id/delete', (req, res, next) => {
-    CourseModel
-        .findByIdAndDelete(req.params.id)
-        .then((course) => {
-            console.log('Course deleted')
-            res.redirect('/courses')
-        })
-        .catch((err) => { next(err) })
+    if (req.app.locals.isTa) {
+        CourseModel
+            .findByIdAndDelete(req.params.id)
+            .then((course) => {
+                console.log('Course deleted')
+                res.redirect('/courses')
+            })
+            .catch((err) => { next(err) })
+    } else {
+        res.render('auth/login', { errorMessage: 'No tenes premiso weon' })
+    }
 })
 
 router.get('/:id/edit', async (req, res, next) => {
-    try {
-        const allstudents = await UserModel.find({ role: STUDENT }).select('username _id')
-        const devs = await UserModel.find({ role: DEV }).select('username _id')
-        const teachers = await UserModel.find({ role: TA }).select('username _id')
-        const course = await CourseModel.findById(req.params.id)
-        const { _id, title, leadTeacher, startDate, endDate, ta, courseImg, description, statusCreate, students } = course
-        res.render('courses/course-edit', { _id, title, leadTeacher, startDate, endDate, ta, courseImg, description, statusCreate, students, allstudents, teachers, devs })
+    if (req.app.locals.isTa) {
+        try {
+            const allstudents = await UserModel.find({ role: STUDENT }).select('username _id')
+            const devs = await UserModel.find({ role: DEV }).select('username _id')
+            const teachers = await UserModel.find({ role: TA }).select('username _id')
+            const course = await CourseModel.findById(req.params.id)
+            const { _id, title, leadTeacher, startDate, endDate, ta, courseImg, description, statusCreate, students } = course
+            res.render('courses/course-edit', { _id, title, leadTeacher, startDate, endDate, ta, courseImg, description, statusCreate, students, allstudents, teachers, devs })
 
-    } catch (err) {
-        next(err)
+        } catch (err) {
+            next(err)
+        }
+    } else {
+        res.render('auth/login', { errorMessage: 'No tenes premiso weon' })
     }
+})
+
+// Check if the student is laready in the course
+
+router.get('/:id/join', (req, res, next) => {
+    CourseModel.updateOne({ _id: req.params.id }, { $addToSet: { students: req.session.currentUser._id } })
+        .then((course) => {
+            console.log("student added to the course")
+            res.redirect('/courses')
+        })
+        .catch((err) => next(err))
 })
 
 router.get('/:id', (req, res, next) => {
@@ -70,9 +92,16 @@ router.post('/:id/edit', (req, res, next) => {
     for (const key in req.body) {
         if (!req.body[key]) delete req.body[key]
     }
+    let status
     const { title, leadTeacher, startDate, endDate, ta, courseImg, description, statusCreate, students } = req.body
+
+    if (statusCreate) {
+        status = 'ON'
+    } else {
+        status = 'OFF'
+    }
     CourseModel
-        .findByIdAndUpdate(req.params.id, { title, leadTeacher, startDate, endDate, ta, courseImg, description, statusCreate, students })
+        .findByIdAndUpdate(req.params.id, { title, leadTeacher, startDate, endDate, ta, courseImg, description, status, students })
         .then((course) => {
             console.log('Course created')
             res.redirect('/courses')
@@ -87,12 +116,10 @@ router.post('/create', (req, res, next) => {
     }
     let status
     const { title, leadTeacher, startDate, endDate, ta, courseImg, description, statusCreate, students } = req.body
-    if (statusCreate === 'on') {
+    if (statusCreate) {
         status = 'ON'
-    } else if (statusCreate === 'off') {
-        status = 'OFF'
     } else {
-        console.log('Course Status error')
+        status = 'OFF'
     }
     CourseModel.create({ title, leadTeacher, startDate, endDate, ta, courseImg, description, status, students })
         .then((course) => {
